+++ classes/Admin/AdminFactory.php
@@ -30,7 +30,7 @@ class AdminFactory implements AC\AdminFactoryInterface {
 	public function create() {
 		return new Admin(
 			new AdminScripts( $this->location ),
-			new PageRequestHandler( $this->storage, $this->location ),
+			new PageRequestHandler( new PageFactory( $this->storage, $this->location ) ),
 			new AdminPageMenuFactory( admin_url( 'options-general.php' ) )
 		);
 	}
