+++ classes/Admin/PageFactoryInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace AC\Admin;
+
+use AC;
+
+interface PageFactoryInterface {
+
+	/**
+	 * @param string $slug
+	 *
+	 * @return Page
+	 */
+	public function create( $slug );
+
+}
