+++ classes/Admin/PageRequestHandler.php
@@ -3,11 +3,6 @@
 namespace AC\Admin;
 
 use AC;
-use AC\Asset\Location;
-use AC\DefaultColumnsRepository;
-use AC\Deprecated\Hooks;
-use AC\Integrations;
-use AC\ListScreenRepository\Storage;
 use AC\Request;
 
 class PageRequestHandler implements AC\PageRequestHandler {
@@ -15,18 +10,12 @@ class PageRequestHandler implements AC\PageRequestHandler {
 	const PARAM_TAB = 'tab';
 
 	/**
-	 * @var Storage
+	 * @var PageFactory
 	 */
-	protected $storage;
+	private $page_factory;
 
-	/**
-	 * @var Location\Absolute
-	 */
-	protected $location;
-
-	public function __construct( Storage $storage, Location\Absolute $location ) {
-		$this->storage = $storage;
-		$this->location = $location;
+	public function __construct( PageFactoryInterface $page_factory ) {
+		$this->page_factory = $page_factory;
 	}
 
 	/**
@@ -35,28 +24,7 @@ class PageRequestHandler implements AC\PageRequestHandler {
 	 * @return Page
 	 */
 	public function handle( Request $request ) {
-
-		switch ( $request->get_query()->get( self::PARAM_TAB ) ) {
-
-			case Page\Help::NAME :
-				return new Page\Help( new Hooks(), $this->location );
-			case Page\Settings::NAME :
-				$sections = new SectionCollection();
-				$sections->add( new Section\General( [ new Section\Partial\ShowEditButton() ] ) )
-				         ->add( new Section\Restore() );
-
-				return new Page\Settings( $sections );
-			case Page\Addons::NAME :
-				return new Page\Addons( $this->location, new Integrations() );
-			default:
-				return new Page\Columns(
-					$this->location,
-					new DefaultColumnsRepository(),
-					new Section\Partial\Menu(),
-					$this->storage,
-					new Preference\ListScreen()
-				);
-		}
+		return $this->page_factory->create( $request->get_query()->get( self::PARAM_TAB ) );
 	}
 
 }
